pyproject.toml
[tool.poetry]
name = "elearning"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "AGPL-3.0-or-later"
readme = "README.md"

homepage = "https://github.com/NC3-LU/eLearning"
repository = "https://github.com/NC3-LU/eLearning"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
psycopg2-binary = "^2.8.4"
cryptography = "^39.0.1"
django-cors-headers = "^3.14.0"
drf-spectacular = "^0.26.1"
drf-spectacular-sidecar = "^2023.4.1"
django-bootstrap5 = "^23.1"
django-otp = "^1.1.6"
django-formtools = "^2.4"
django-two-factor-auth = "^1.15.2"
phonenumbers = "^8.13.11"
django-import-export = "^3.2.0"
parler = "^1.0.1"
django-parler = "^2.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
pre-commit = "^2.20.0"
black = "^22.10.0"
pyupgrade = "^3.2.2"
autoflake = "^1.7.7"
mypy = "^0.991"
types-requests = "^2.28.11.5"
pip-audit = "^2.5.4"
isort = "^5.12.0"
django-debug-toolbar = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
